MeasurementCalibrator uses device calibration to calibrate input measurement

+++ src/main/kotlin/com/noicesoftware/calibre/calibration/Equal.kt
@@ -1,5 +1,7 @@
 package com.noicesoftware.calibre.calibration
 
+import java.math.BigDecimal
+
 class Equal : IEquation {
-    override fun calculate(value: Double) = value
+    override fun calculate(value: BigDecimal) = value
 }
