+++ src/main/kotlin/com/noicesoftware/calibre/calibration/Linear.kt
@@ -1,5 +1,7 @@
 package com.noicesoftware.calibre.calibration
 
-class Linear(private val gradient: Double, private val offset: Double) : IEquation {
-    override fun calculate(value: Double) = gradient * value + offset
+import java.math.BigDecimal
+
+class Linear(private val gradient: BigDecimal, private val offset: BigDecimal) : IEquation {
+    override fun calculate(value: BigDecimal) = gradient * value + offset
 }
